+++ NEWS.md
@@ -26,6 +26,7 @@ in the augment method for the chi sq test, .residuals column was renamed to .res
 - Added new vignette detailing use of `modelgenerics` and `modeltests` packages
 - Added `data` argument to `augment()` generic
 - tidy.kmeans now defaults to using variable names in output columns
+- `tidy_optim()` provides the standard error if the Hessian is present.
 
 ## Deprecations